requirements.txt
pymumble==1.4
python-valve==0.2.1
boto3==1.16.19
botocore==1.19.19